Consolidate core to single file
